+++ frontend/techpick-extension/src/widgets/TagInput/index.tsx
@@ -1,71 +1,23 @@
-import { useEffect, useRef, useState } from 'react';
-import { Command } from 'cmdk';
+import { useRef, useState } from 'react';
 import {
-  SelectedTagContent,
-  SelectedTagLayout,
   useTagStore,
+  SelectedTagItem,
   SelectedTagListLayout,
 } from '@/entities/tag';
-import { DeselectTagButton, DeleteTagDialog } from '@/features/tag';
-import { TagInfoEditPopoverButton } from '@/widgets/TagInfoEditPopoverButton';
-import {
-  tagInputStyle,
-  tagDialogTrigger,
-  tagDialogPortal,
-  selectedTagItemStyle,
-  commandInputStyle,
-  commandInputLayout,
-} from './TagInput.css';
-import { filterCommandItems } from './TagInput.lib';
+import { TagSelectionDialog } from '@/widgets/TagSelectionDialog';
+
+import { tagInputStyle, tagDialogTrigger } from './TagInput.css';
 
 export function TagInput() {
   const [open, setOpen] = useState(false);
-  const [tagInputValue, setTagInputValue] = useState('');
-  const [canCreateTag, setCanCreateTag] = useState(false);
-  const tagInputContainerRef = useRef<HTMLDivElement | null>(null);
-  const tagInputRef = useRef<HTMLInputElement | null>(null);
-  const {
-    tagList,
-    selectedTagList,
-    fetchingTagState,
-    selectTag,
-    fetchingTagList,
-    createTag,
-  } = useTagStore();
+
+  const tagInputContainerRef = useRef<HTMLDivElement>(null);
+
+  const { selectedTagList } = useTagStore();
 
   const openDialog = () => {
     setOpen(true);
   };
-
-  const focusTagInput = () => {
-    tagInputRef.current?.focus();
-  };
-
-  const clearTagInputValue = () => {
-    setTagInputValue('');
-  };
-
-  useEffect(
-    function fetchTagList() {
-      fetchingTagList();
-    },
-    [fetchingTagList]
-  );
-
-  useEffect(
-    function checkIsCreatable() {
-      const isUnique = !tagList.some((tag) => tag.name === tagInputValue);
-      const isNotInitialValue = tagInputValue.trim() !== '';
-      const isCreatable = isUnique && isNotInitialValue;
-
-      setCanCreateTag(isCreatable);
-    },
-    [tagInputValue, tagList]
-  );
-
-  if (fetchingTagState.isPending) {
-    return <h1>Loading...</h1>;
-  }
 
   return (
     <div
@@ -81,82 +33,16 @@ export function TagInput() {
       >
         <SelectedTagListLayout>
           {selectedTagList.map((tag) => (
-            <SelectedTagLayout key={tag.id}>
-              <SelectedTagContent>{tag.name}</SelectedTagContent>
-            </SelectedTagLayout>
+            <SelectedTagItem key={tag.id} tag={tag} />
           ))}
         </SelectedTagListLayout>
       </div>
 
-      <Command.Dialog
+      <TagSelectionDialog
         open={open}
         onOpenChange={setOpen}
-        container={tagInputContainerRef.current ?? undefined}
-        className={tagDialogPortal}
-        filter={filterCommandItems}
-      >
-        {/**선택한 태그 리스트 */}
-        <SelectedTagListLayout>
-          {selectedTagList.map((tag) => (
-            <SelectedTagLayout key={tag.id}>
-              <SelectedTagContent>{tag.name}</SelectedTagContent>
-              <DeselectTagButton tag={tag} onClick={focusTagInput} />
-            </SelectedTagLayout>
-          ))}
-
-          <div className={commandInputLayout}>
-            <Command.Input
-              className={commandInputStyle}
-              ref={tagInputRef}
-              value={tagInputValue}
-              onValueChange={setTagInputValue}
-            />
-          </div>
-        </SelectedTagListLayout>
-
-        {/**전체 태그 리스트 */}
-        <Command.List>
-          <Command.Empty>No results found.</Command.Empty>
-
-          {tagList.map((tag) => (
-            <Command.Item
-              key={tag.id}
-              className={selectedTagItemStyle}
-              onSelect={() => {
-                selectTag(tag);
-                focusTagInput();
-                clearTagInputValue();
-              }}
-              keywords={[tag.name]}
-            >
-              {tag.name}
-              <TagInfoEditPopoverButton tag={tag} />
-            </Command.Item>
-          ))}
-
-          {canCreateTag && (
-            <Command.Item
-              className={selectedTagItemStyle}
-              value={tagInputValue}
-              keywords={['생성']}
-              onSelect={async () => {
-                const newTag = await createTag(tagInputValue);
-
-                if (newTag) {
-                  selectTag(newTag);
-                  focusTagInput();
-                  clearTagInputValue();
-                }
-              }}
-              disabled={!canCreateTag}
-            >
-              {tagInputValue} 생성
-            </Command.Item>
-          )}
-        </Command.List>
-
-        <DeleteTagDialog />
-      </Command.Dialog>
+        container={tagInputContainerRef}
+      />
     </div>
   );
 }
